requirements.txt
asgiref==3.5.1
Django==4.0.4
django-filter==21.1
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2
